+++ test/resources/checkers/c_c++-warning.c
@@ -5,4 +5,5 @@ int f(int x)
      int unused;
      unsigned int y = 10;
      return x < y ? ++x : x;
+     #warning
 }